requirements.txt
aiohttp==3.10.8
colorama==0.4.6
Pillow==10.4.0
PyPDF2==3.0.1
python-dotenv==1.0.1